pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voiceinput"
version = "0.1.0"
description = "A macOS terminal voice input tool built on top of funasr."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "voiceinput contributors" }
]
keywords = ["asr", "funasr", "voice", "cli", "speech-to-text"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
  "Topic :: Software Development :: User Interfaces",
]
dependencies = [
  "numpy>=1.26.0",
  "funasr==1.3.1",
  "torch>=2.0.0",
  "sounddevice>=0.4.0",
  "scipy>=1.10.0",
  "soundfile>=0.12.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
voice = "voice_entry:main"

[tool.setuptools]
include-package-data = true
py-modules = [
  "app_factory",
  "asr_engine",
  "audio_processor",
  "bootstrap",
  "cli",
  "recorder",
  "recording_session",
  "runtime_ui",
  "text_processing",
  "voice_entry",
]

[tool.setuptools.data-files]
"config" = ["config/settings.example.json"]

[tool.setuptools.package-dir]
"" = "src"
